Revert "basic components and navigation"
