+++ package.json
@@ -9,20 +9,13 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/webpack-config": "^18.0.1",
-    "@react-native-community/masked-view": "^0.1.11",
-    "@react-navigation/native": "^6.1.6",
-    "@react-navigation/native-stack": "^6.9.12",
     "expo": "~48.0.5",
     "expo-status-bar": "~1.4.4",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
     "react-native": "0.71.3",
-    "react-native-gesture-handler": "^2.9.0",
-    "react-native-reanimated": "^3.0.0",
-    "react-native-safe-area-context": "^4.5.0",
-    "react-native-screens": "~3.20.0",
-    "react-native-web": "~0.18.10"
+    "react-native-web": "~0.18.10",
+    "react-dom": "18.2.0",
+    "@expo/webpack-config": "^18.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
