--- src/components/login.jsx
@@ -1,30 +1,0 @@
-import React, { useState } from 'react';
-import { View, TextInput, StyleSheet, Button, Text } from 'react-native';
-import { useNavigation } from '@react-navigation/native';
-
-const Login = () => {
-    const navigation = useNavigation();
-    const handleNav = () => {
-      navigation.navigate('Dashboard');
-    }
-
-    return (
-      <View style={styles.container}>
-        <Text>Login goes here</Text>
-        <TextInput></TextInput>
-        <Button title="Login" onPress={handleNav}/>
-      </View>
-    );
-}
-
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
-
-export default Login;
