--- src/components/saved.jsx
@@ -1,23 +1,0 @@
-import React, { useState } from 'react';
-import { View, TextInput, StyleSheet, Button, Text } from 'react-native';
-import { useNavigation } from '@react-navigation/native';
-
-const Saved = () => {
-    return (
-      <View style={styles.container}>
-        <Text>Saved</Text>
-      </View>
-    );
-}
-
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
-
-export default Saved;
